mcp_surf_server/requirements.txt
mcp>=1.2.0
fastapi
uvicorn[standard]
httpx
python-dotenv 